requirements.txt
torch
transformers
pandas
imaplib2
python-dotenv